requirements.txt
langgraph>=0.1.6
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0

docx2txt>=0.8

requests>=2.31.0
python-dotenv>=1.0.1
typing-extensions>=4.8.0
streamlit>=1.35.0